requirements.txt
python-telegram-bot==21.6
supabase>=2.28.0
python-dotenv==1.0.1
